Group compatibility suites by upstream package
